Register litedb adapter for Rails >= 7.2
Rails main branch is failing to find the litedb adapter because some changes introduced here: https://github.com/rails/rails/commit/009c7e74117690f0dbe200188a929b345c9306c1

This commit registers the adapter. There may be better ways to do it, this is a quick patch to make it work.

Another way is to do add the new line in an initializer that can be added by the `install` generator.